Get repo visibility from GitHub
Use the GitHub API to determine repo visibility.  Remove org argument
to functions, derive it using GitHubRepo.  RestartBuildFunction still
takes a parameter indicating which Travis API to hit since it does not
get repo owner and name but a build ID, so there is no way to derive
the Travis API.  We did change the parameter to accept "public" and
"private".

Move retry logic out into own object so it can be used by both Travis
and GitHub methods.

Update rug to 1.0.0-m.3.

Make RepoSlug type because the owner and name were almost always used
together and with a type we can avoid putting multiple string
arguments in the wrong order.

Make the GitHubRepo trait simpler, preferring to pass in arguments to
functions rather than storing state in the object.

Make types for tokens so they do not get mixed up